test mmap loading

+++ src/net.h
@@ -98,14 +98,14 @@ public:
     // load network structure from external memory
     // memory pointer must be 32-bit aligned
     // return bytes consumed
-    int load_param(const unsigned char* mem);
+    size_t load_param(const unsigned char* mem);
 
     // reference network weight data from external memory
     // weight data is not copied but referenced
     // so external memory should be retained when used
     // memory pointer must be 32-bit aligned
     // return bytes consumed
-    int load_model(const unsigned char* mem);
+    size_t load_model(const unsigned char* mem);
 
 #if NCNN_PLATFORM_API
 #if __ANDROID_API__ >= 9
